+++ src/gui/item_models/property_model_single.hpp
@@ -18,6 +18,7 @@ public:
     enum Columns
     {
         ColumnName,
+        ColumnToggleKeyframe,
         ColumnValue,
 
         ColumnCount